+++ src/test/java/it/tredi/msa/test/EmlExtractionTest.java
@@ -4,14 +4,8 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
 import java.io.File;
-import java.io.IOException;
 import java.util.List;
 
-import javax.mail.Message;
-import javax.mail.MessagingException;
-import javax.mail.internet.MimeMessage;
-
-import org.apache.commons.io.FileUtils;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -26,24 +20,8 @@ import it.tredi.msa.test.conf.MsaTesterApplication;
  */
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringBootTest(classes = MsaTesterApplication.class)
-public class EmlExtractionTest {
+public class EmlExtractionTest extends EmlReader {
 	
-	private static final String EML_LOCATION = "eml";
-	
-	/**
-	 * Creazione dell'oggetto message a partire da un file salvato su disco
-	 * @param file
-	 * @return
-	 * @throws IOException 
-	 * @throws MessagingException 
-	 */
-	private Message _readEmlFile(File file) throws MessagingException, IOException {
-		Message message = null;
-		if (file != null && file.exists())
-	        message = new MimeMessage(null, FileUtils.openInputStream(file));
-		return message;
-	}
-
 	/**
 	 * Test di estrazione di un messaggio contenente una parte NULL
 	 */
@@ -54,7 +32,7 @@ public class EmlExtractionTest {
 		
 		System.out.println("input file = " + fileName);
 		
-		ParsedMessage parsed = new ParsedMessage(_readEmlFile(file));
+		ParsedMessage parsed = new ParsedMessage(readEmlFile(file));
 		
 		assertNotNull(parsed);
 		assertNotNull(parsed.getMessageId());
@@ -89,7 +67,7 @@ public class EmlExtractionTest {
 		
 		System.out.println("input file = " + fileName);
 		
-		ParsedMessage parsed = new ParsedMessage(_readEmlFile(file));
+		ParsedMessage parsed = new ParsedMessage(readEmlFile(file));
 		
 		assertNotNull(parsed);
 		assertNotNull(parsed.getMessageId());
@@ -110,6 +88,34 @@ public class EmlExtractionTest {
 		
 		assertNotNull(fromDatiCert);
 		assertEquals("paolapenna@ordineavvocatiroma.org", fromDatiCert);
+	}
+	
+	/**
+	 * Errore NotWellFormed su ServerCommand.send su eXtraWay server
+	 * @throws Exception
+	 */
+	@Test
+	public void notWellFormed2Extraction() throws Exception {
+		String fileName = "notWellFormed_2.eml";
+		File file = ResourceUtils.getFile("classpath:" + EML_LOCATION + "/" + fileName);
+		
+		System.out.println("input file = " + fileName);
+		
+		ParsedMessage parsed = new ParsedMessage(readEmlFile(file));
+		
+		assertNotNull(parsed);
+		assertNotNull(parsed.getMessageId());
+		
+		System.out.println("messageId = " + parsed.getMessageId());
+		System.out.println("subject = " + parsed.getSubject());
+		System.out.println("from address = " + parsed.getFromAddress());
+		
+		List<String> attachments = parsed.getAttachmentsName();
+		System.out.println("attachments count = " + attachments.size());
+		for (String name : attachments)
+			System.out.println("\tattach name = " + name);
+		
+		assertEquals(5, parsed.getAttachments().size());
 	}
 	
 }